chore: refactor intro cog to new approach
