+++ koala/cogs/__init__.py
@@ -3,3 +3,4 @@ from .base import BaseCog
 from .announce import Announce
 from .twitch_alert import TwitchAlert
 from .colour_role import ColourRole
+from .intro_cog import IntroCog
