+++ koala/db.py
@@ -18,7 +18,7 @@ from sqlalchemy.orm import sessionmaker
 
 
 # Own modules
-from koala.models import mapper_registry, KoalaExtensions, GuildExtensions, GuildWelcomeMessages
+from koala.models import mapper_registry, KoalaExtensions, GuildExtensions
 from koala.env import DB_KEY, ENCRYPTED_DB
 from koala.utils.KoalaUtils import get_arg_config_path, format_config_path
 
@@ -214,63 +214,3 @@ def clear_all_tables(tables):
         for table in tables:
             session.execute('DELETE FROM ' + table + ';')
             session.commit()
-
-
-def fetch_guild_welcome_message(guild_id):
-    """
-    Fetches the guild welcome message for a given guild
-
-    :param guild_id: Discord guild ID for a given server
-    """
-    with session_manager() as session:
-        msg = session.execute(select(GuildWelcomeMessages.welcome_message)
-                              .where(GuildWelcomeMessages.guild_id == guild_id)).one_or_none()
-    if not msg:
-        return None
-    else:
-        return msg.welcome_message
-
-
-def update_guild_welcome_message(guild_id, new_message: str):
-    """
-    Update guild welcome message for a given guild
-
-    :param guild_id: Discord guild ID for a given server
-    :param new_message: The new guild welcome message to be set
-    """
-    with session_manager() as session:
-        session.execute(update(GuildWelcomeMessages)
-                        .where(GuildWelcomeMessages.guild_id == guild_id)
-                        .values(welcome_message=new_message))
-        session.commit()
-    return new_message
-
-
-def remove_guild_welcome_message(guild_id):
-    """
-    Removes the guild welcome message from a given guild
-
-    :param guild_id: Discord guild ID for a given server
-    """
-    with session_manager() as session:
-        welcome_message = session.execute(select(GuildWelcomeMessages).filter_by(guild_id=guild_id))\
-            .scalars().one_or_none()
-        if welcome_message:
-            session.delete(welcome_message)
-            session.commit()
-            return 1
-        return 0
-
-
-def new_guild_welcome_message(guild_id):
-    """
-    Sets the default guild welcome message to a given guild
-
-    :param guild_id: Discord guild ID for a given server
-    """
-    from koala.cogs.IntroCog import DEFAULT_WELCOME_MESSAGE
-
-    with session_manager() as session:
-        session.add(GuildWelcomeMessages(guild_id=guild_id, welcome_message=DEFAULT_WELCOME_MESSAGE))
-        session.commit()
-    return fetch_guild_welcome_message(guild_id)
