+++ koala/models.py
@@ -34,15 +34,3 @@ class GuildExtensions:
     def __repr__(self):
         return "GuildExtensions(%s, %s)>" % \
                (self.extension_id, self.guild_id)
-
-
-@mapper_registry.mapped
-class GuildWelcomeMessages:
-    __tablename__ = 'GuildWelcomeMessages'
-    guild_id = Column(Integer, primary_key=True)
-    welcome_message = Column(String, nullable=True)
-
-    def __repr__(self):
-        return "GuildWelcomeMessages(%s, %s)>" % \
-               (self.guild_id, self.welcome_message)
-
