+++ koala/utils/KoalaDBManager.py
@@ -218,47 +218,3 @@ class KoalaDBManager:
         """
         db.clear_all_tables(tables)
 
-    def fetch_guild_welcome_message(self, guild_id):
-        """
-        Fetches the guild welcome message for a given guild
-
-        :param guild_id: Discord guild ID for a given server
-
-        .. deprecated:: 0.4.5
-           This method is now found in base_db
-        """
-        return db.fetch_guild_welcome_message(guild_id)
-
-    def update_guild_welcome_message(self, guild_id, new_message: str):
-        """
-        Update guild welcome message for a given guild
-
-        :param guild_id: Discord guild ID for a given server
-        :param new_message: The new guild welcome message to be set
-
-        .. deprecated:: 0.4.5
-           This method is now found in base_db
-        """
-        return db.update_guild_welcome_message(guild_id, new_message)
-
-    def remove_guild_welcome_message(self, guild_id):
-        """
-        Removes the guild welcome message from a given guild
-
-        :param guild_id: Discord guild ID for a given server
-
-        .. deprecated:: 0.4.5
-           This method is now found in base_db
-        """
-        return db.remove_guild_welcome_message(guild_id)
-
-    def new_guild_welcome_message(self, guild_id):
-        """
-        Sets the default guild welcome message to a given guild
-
-        :param guild_id: Discord guild ID for a given server
-
-        .. deprecated:: 0.4.5
-           This method is now found in base_db
-        """
-        return db.new_guild_welcome_message(guild_id)
